gem5: Drop support for ancient gem5 versions
Support for older gem5 versions involved checking the gem5 revisions using
mercurial commands. This is problematic now that gem5 is git based with a
mercurial read-only mirror. When the check fails or returns an incorrect
value it will cause mysterious build errors trying to use old gem5 APIs.
Additionally, the latest version of gem5 is generally stable due to rigorous
review processes and memory APIs have not changed in quite some time, so
assuming the most recent version of gem5 is reasonable.